Switch to webpack
