--- dev/JSLintNet.json
@@ -1,5 +1,0 @@
-﻿{
-  "ignore": [
-    "\\bower_components\\"
-  ]
-}
