--- dev/bower.json
@@ -1,27 +1,0 @@
-{
-  "name": "jump-n-bump",
-  "version": "0.1.0",
-  "description": "An instant real-time multiplayer-only keyboard-crowding classic, this game pits up to four human-operated cute fuzzy widdle bunnies against each other for one goal only: lagomorphic cranial domination. Originally created by Brainchild Design",
-  "main": "jump-n-bump.js",
-  "moduleType": [
-    "node"
-  ],
-  "keywords": [
-    "game"
-  ],
-  "authors": [
-    "grahamthecoder"
-  ],
-  "license": "GNU GENERAL PUBLIC LICENSE",
-  "homepage": "https://github.com/GrahamTheCoder/jump-n-bump",
-  "ignore": [
-    "**/.*",
-    "node_modules",
-    "bower_components",
-    "test",
-    "tests"
-  ],
-  "dependencies": {
-    "knockoutjs": "~3.3.0"
-  }
-}
