--- dev/build.bat
@@ -1,3 +1,0 @@
-@ECHO OFF
-
-node_modules\.bin\grunt
