--- dev/first-time-setup.bat
@@ -1,8 +1,0 @@
-@ECHO OFF
-ECHO Note: Script requires NPM on the PATH
-ECHO If successful, run 'build.bat' to build
-call npm install
-node_modules\.bin\bower install
-
-
-PAUSE
