--- dev/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "jump-n-bump",
-  "version": "1.0.1",
-  "description": "HTML5 port of the open source DOS classic by Brainchild Design",
-  "repository": "https://github.com/GrahamTheCoder/jump-n-bump",
-  "devDependencies": {
-    "bower": "*",
-    "grunt": "*",
-    "grunt-bower": "*",
-    "grunt-bower-concat": "*",
-    "grunt-cli": "*",
-    "grunt-contrib-concat": "*",
-    "grunt-contrib-copy": "*",
-    "grunt-contrib-uglify": "*",
-    "grunt-contrib-watch": "*",
-    "grunt-zip": "*",
-    "main-bower-files": "*",
-    "nogit": "*"
-  }
-}
