--- jump-n-bump.sln
@@ -1,27 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.22823.1
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "jump-n-bump", "dev\jump-n-bump.csproj", "{A8333A5E-A335-4B9E-9631-75B1F8AD21CF}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{B65FDFBA-409C-4CCE-985A-3A388768970C}"
-	ProjectSection(SolutionItems) = preProject
-		README.md = README.md
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A8333A5E-A335-4B9E-9631-75B1F8AD21CF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A8333A5E-A335-4B9E-9631-75B1F8AD21CF}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A8333A5E-A335-4B9E-9631-75B1F8AD21CF}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A8333A5E-A335-4B9E-9631-75B1F8AD21CF}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
