+++ package.json
@@ -1,0 +1,24 @@
+{
+  "name": "jump-n-bump",
+  "version": "1.0.1",
+  "description": "HTML5 port of the open source DOS classic by Brainchild Design",
+  "repository": "https://github.com/GrahamTheCoder/jump-n-bump",
+  "devDependencies": {
+    "babel-core": "^6.26.0",
+    "babel-loader": "^7.1.2",
+    "babel-preset-env": "^1.6.0",
+    "css-loader": "^0.28.7",
+    "extract-text-webpack-plugin": "^3.0.1",
+    "file-loader": "^1.1.5",
+    "html-webpack-plugin": "^2.30.1",
+    "webpack": "^3.6.0",
+    "webpack-dev-server": "^2.9.1"
+  },
+  "scripts": {
+    "dev": "webpack-dev-server",
+    "build": "webpack"
+  },
+  "dependencies": {
+    "knockout": "^3.3.0"
+  }
+}
