+++ src/asset_data/object_gobs.js
@@ -1,0 +1,563 @@
+export const object_gobs = [
+{
+    x: 0,
+    y: 0,
+    width: 17,
+    height: 7,
+    hotspot_x: 0,
+    hotspot_y: 3
+}, {
+    x: 32,
+    y: 0,
+    width: 17,
+    height: 8,
+    hotspot_x: 0,
+    hotspot_y: 4
+}, {
+    x: 64,
+    y: 0,
+    width: 17,
+    height: 12,
+    hotspot_x: 0,
+    hotspot_y: 8
+}, {
+    x: 96,
+    y: 0,
+    width: 17,
+    height: 10,
+    hotspot_x: 0,
+    hotspot_y: 6
+}, {
+    x: 128,
+    y: 0,
+    width: 17,
+    height: 8,
+    hotspot_x: 0,
+    hotspot_y: 4
+}, {
+    x: 160,
+    y: 0,
+    width: 17,
+    height: 7,
+    hotspot_x: 0,
+    hotspot_y: 3
+}, {
+    x: 192,
+    y: 0,
+    width: 12,
+    height: 6,
+    hotspot_x: 5,
+    hotspot_y: 5
+}, {
+    x: 224,
+    y: 0,
+    width: 16,
+    height: 8,
+    hotspot_x: 7,
+    hotspot_y: 7
+}, {
+    x: 256,
+    y: 0,
+    width: 20,
+    height: 10,
+    hotspot_x: 9,
+    hotspot_y: 9
+}, {
+    x: 288,
+    y: 0,
+    width: 23,
+    height: 11,
+    hotspot_x: 11,
+    hotspot_y: 10
+}, {
+    x: 320,
+    y: 0,
+    width: 25,
+    height: 12,
+    hotspot_x: 12,
+    hotspot_y: 11
+}, {
+    x: 352,
+    y: 0,
+    width: 27,
+    height: 11,
+    hotspot_x: 13,
+    hotspot_y: 10
+}, {
+    x: 0,
+    y: 14,
+    width: 29,
+    height: 10,
+    hotspot_x: 14,
+    hotspot_y: 8
+}, {
+    x: 32,
+    y: 14,
+    width: 30,
+    height: 7,
+    hotspot_x: 15,
+    hotspot_y: 5
+}, {
+    x: 64,
+    y: 14,
+    width: 29,
+    height: 4,
+    hotspot_x: 16,
+    hotspot_y: 2
+}, {
+    x: 96,
+    y: 14,
+    width: 2,
+    height: 2,
+    hotspot_x: 0,
+    hotspot_y: 0
+}, {
+    x: 128,
+    y: 14,
+    width: 4,
+    height: 4,
+    hotspot_x: 1,
+    hotspot_y: 1
+}, {
+    x: 160,
+    y: 14,
+    width: 5,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 2
+}, {
+    x: 192,
+    y: 14,
+    width: 5,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 2
+}, {
+    x: 224,
+    y: 14,
+    width: 5,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 2
+}, {
+    x: 256,
+    y: 14,
+    width: 7,
+    height: 5,
+    hotspot_x: 4,
+    hotspot_y: 4
+}, {
+    x: 288,
+    y: 14,
+    width: 8,
+    height: 4,
+    hotspot_x: 5,
+    hotspot_y: 3
+}, {
+    x: 320,
+    y: 14,
+    width: 9,
+    height: 3,
+    hotspot_x: 5,
+    hotspot_y: 2
+}, {
+    x: 352,
+    y: 14,
+    width: 8,
+    height: 3,
+    hotspot_x: 5,
+    hotspot_y: 1
+}, {
+    x: 0,
+    y: 28,
+    width: 8,
+    height: 4,
+    hotspot_x: 5,
+    hotspot_y: 1
+}, {
+    x: 32,
+    y: 28,
+    width: 7,
+    height: 5,
+    hotspot_x: 4,
+    hotspot_y: 1
+}, {
+    x: 64,
+    y: 28,
+    width: 7,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 4
+}, {
+    x: 96,
+    y: 28,
+    width: 8,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 3
+}, {
+    x: 128,
+    y: 28,
+    width: 9,
+    height: 3,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 160,
+    y: 28,
+    width: 8,
+    height: 3,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 192,
+    y: 28,
+    width: 8,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 224,
+    y: 28,
+    width: 7,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 256,
+    y: 28,
+    width: 7,
+    height: 5,
+    hotspot_x: 4,
+    hotspot_y: 4
+}, {
+    x: 288,
+    y: 28,
+    width: 8,
+    height: 4,
+    hotspot_x: 5,
+    hotspot_y: 3
+}, {
+    x: 320,
+    y: 28,
+    width: 9,
+    height: 3,
+    hotspot_x: 5,
+    hotspot_y: 2
+}, {
+    x: 352,
+    y: 28,
+    width: 8,
+    height: 3,
+    hotspot_x: 5,
+    hotspot_y: 1
+}, {
+    x: 0,
+    y: 42,
+    width: 8,
+    height: 4,
+    hotspot_x: 5,
+    hotspot_y: 1
+}, {
+    x: 32,
+    y: 42,
+    width: 7,
+    height: 5,
+    hotspot_x: 4,
+    hotspot_y: 1
+}, {
+    x: 64,
+    y: 42,
+    width: 7,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 4
+}, {
+    x: 96,
+    y: 42,
+    width: 8,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 3
+}, {
+    x: 128,
+    y: 42,
+    width: 9,
+    height: 3,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 160,
+    y: 42,
+    width: 8,
+    height: 3,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 192,
+    y: 42,
+    width: 8,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 224,
+    y: 42,
+    width: 7,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 256,
+    y: 42,
+    width: 6,
+    height: 4,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 288,
+    y: 42,
+    width: 6,
+    height: 5,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 320,
+    y: 42,
+    width: 4,
+    height: 6,
+    hotspot_x: 2,
+    hotspot_y: 3
+}, {
+    x: 352,
+    y: 42,
+    width: 5,
+    height: 6,
+    hotspot_x: 2,
+    hotspot_y: 3
+}, {
+    x: 0,
+    y: 56,
+    width: 6,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 2
+}, {
+    x: 32,
+    y: 56,
+    width: 6,
+    height: 5,
+    hotspot_x: 2,
+    hotspot_y: 2
+}, {
+    x: 64,
+    y: 56,
+    width: 4,
+    height: 6,
+    hotspot_x: 1,
+    hotspot_y: 2
+}, {
+    x: 96,
+    y: 56,
+    width: 5,
+    height: 6,
+    hotspot_x: 2,
+    hotspot_y: 3
+}, {
+    x: 128,
+    y: 56,
+    width: 6,
+    height: 4,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 160,
+    y: 56,
+    width: 6,
+    height: 5,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 192,
+    y: 56,
+    width: 4,
+    height: 6,
+    hotspot_x: 2,
+    hotspot_y: 3
+}, {
+    x: 224,
+    y: 56,
+    width: 5,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 256,
+    y: 56,
+    width: 6,
+    height: 4,
+    hotspot_x: 4,
+    hotspot_y: 2
+}, {
+    x: 288,
+    y: 56,
+    width: 6,
+    height: 5,
+    hotspot_x: 4,
+    hotspot_y: 3
+}, {
+    x: 320,
+    y: 56,
+    width: 4,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 352,
+    y: 56,
+    width: 5,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 0,
+    y: 70,
+    width: 6,
+    height: 4,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 32,
+    y: 70,
+    width: 6,
+    height: 5,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 64,
+    y: 70,
+    width: 4,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 96,
+    y: 70,
+    width: 5,
+    height: 6,
+    hotspot_x: 4,
+    hotspot_y: 2
+}, {
+    x: 128,
+    y: 70,
+    width: 6,
+    height: 4,
+    hotspot_x: 5,
+    hotspot_y: 1
+}, {
+    x: 160,
+    y: 70,
+    width: 6,
+    height: 5,
+    hotspot_x: 5,
+    hotspot_y: 2
+}, {
+    x: 192,
+    y: 70,
+    width: 4,
+    height: 6,
+    hotspot_x: 4,
+    hotspot_y: 3
+}, {
+    x: 224,
+    y: 70,
+    width: 5,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 256,
+    y: 70,
+    width: 6,
+    height: 4,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 288,
+    y: 70,
+    width: 6,
+    height: 5,
+    hotspot_x: 3,
+    hotspot_y: 2
+}, {
+    x: 320,
+    y: 70,
+    width: 4,
+    height: 6,
+    hotspot_x: 2,
+    hotspot_y: 2
+}, {
+    x: 352,
+    y: 70,
+    width: 5,
+    height: 6,
+    hotspot_x: 4,
+    hotspot_y: 2
+}, {
+    x: 0,
+    y: 84,
+    width: 6,
+    height: 4,
+    hotspot_x: 5,
+    hotspot_y: 1
+}, {
+    x: 32,
+    y: 84,
+    width: 6,
+    height: 5,
+    hotspot_x: 5,
+    hotspot_y: 2
+}, {
+    x: 64,
+    y: 84,
+    width: 4,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 96,
+    y: 84,
+    width: 5,
+    height: 6,
+    hotspot_x: 3,
+    hotspot_y: 3
+}, {
+    x: 128,
+    y: 84,
+    width: 5,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 160,
+    y: 84,
+    width: 4,
+    height: 4,
+    hotspot_x: 2,
+    hotspot_y: 1
+}, {
+    x: 192,
+    y: 84,
+    width: 2,
+    height: 3,
+    hotspot_x: 1,
+    hotspot_y: 1
+}, {
+    x: 224,
+    y: 84,
+    width: 1,
+    height: 1,
+    hotspot_x: 0,
+    hotspot_y: 0
+}
+];
