+++ src/asset_data/rabbit_gobs.js
@@ -1,0 +1,578 @@
+export const rabbit_gobs = [
+{
+    x: 0,
+    y: 0,
+    width: 13,
+    height: 15,
+    hotspot_x: -2,
+    hotspot_y: -1
+},
+{
+    x: 19,
+    y: 0,
+    width: 14,
+    height: 14,
+    hotspot_x: -2,
+    hotspot_y: 0
+},
+{
+    x: 38,
+    y: 0,
+    width: 14,
+    height: 16,
+    hotspot_x: -2,
+    hotspot_y: 2
+},
+{
+    x: 57,
+    y: 0,
+    width: 13,
+    height: 16,
+    hotspot_x: -3,
+    hotspot_y: 2
+},
+{
+    x: 76,
+    y: 0,
+    width: 14,
+    height: 14,
+    hotspot_x: -3,
+    hotspot_y: -2
+},
+{
+    x: 95,
+    y: 0,
+    width: 15,
+    height: 14,
+    hotspot_x: -1,
+    hotspot_y: 0
+},
+{
+    x: 114,
+    y: 0,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 133,
+    y: 0,
+    width: 17,
+    height: 16,
+    hotspot_x: 1,
+    hotspot_y: 0
+},
+{
+    x: 152,
+    y: 0,
+    width: 17,
+    height: 13,
+    hotspot_x: -1,
+    hotspot_y: -3
+},
+{
+    x: 171,
+    y: 0,
+    width: 13,
+    height: 15,
+    hotspot_x: -1,
+    hotspot_y: -1
+},
+{
+    x: 190,
+    y: 0,
+    width: 14,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 209,
+    y: 0,
+    width: 14,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 228,
+    y: 0,
+    width: 13,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 247,
+    y: 0,
+    width: 14,
+    height: 14,
+    hotspot_x: 1,
+    hotspot_y: -1
+},
+{
+    x: 266,
+    y: 0,
+    width: 15,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 1
+},
+{
+    x: 285,
+    y: 0,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 304,
+    y: 0,
+    width: 17,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 323,
+    y: 0,
+    width: 17,
+    height: 13,
+    hotspot_x: 2,
+    hotspot_y: -4
+},
+{
+    x: 342,
+    y: 0,
+    width: 13,
+    height: 15,
+    hotspot_x: -2,
+    hotspot_y: -1
+},
+{
+    x: 361,
+    y: 0,
+    width: 14,
+    height: 14,
+    hotspot_x: -2,
+    hotspot_y: 0
+},
+{
+    x: 380,
+    y: 0,
+    width: 14,
+    height: 16,
+    hotspot_x: -2,
+    hotspot_y: 2
+},
+{
+    x: 0,
+    y: 18,
+    width: 13,
+    height: 16,
+    hotspot_x: -3,
+    hotspot_y: 2
+},
+{
+    x: 19,
+    y: 18,
+    width: 14,
+    height: 14,
+    hotspot_x: -3,
+    hotspot_y: -2
+},
+{
+    x: 38,
+    y: 18,
+    width: 15,
+    height: 14,
+    hotspot_x: -1,
+    hotspot_y: 0
+},
+{
+    x: 57,
+    y: 18,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 76,
+    y: 18,
+    width: 17,
+    height: 16,
+    hotspot_x: 1,
+    hotspot_y: 0
+},
+{
+    x: 95,
+    y: 18,
+    width: 17,
+    height: 13,
+    hotspot_x: -1,
+    hotspot_y: -3
+},
+{
+    x: 114,
+    y: 18,
+    width: 13,
+    height: 15,
+    hotspot_x: -1,
+    hotspot_y: -1
+},
+{
+    x: 133,
+    y: 18,
+    width: 14,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 152,
+    y: 18,
+    width: 14,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 171,
+    y: 18,
+    width: 13,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 190,
+    y: 18,
+    width: 14,
+    height: 14,
+    hotspot_x: 1,
+    hotspot_y: -1
+},
+{
+    x: 209,
+    y: 18,
+    width: 15,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 1
+},
+{
+    x: 228,
+    y: 18,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 247,
+    y: 18,
+    width: 17,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 266,
+    y: 18,
+    width: 17,
+    height: 13,
+    hotspot_x: 2,
+    hotspot_y: -4
+},
+{
+    x: 285,
+    y: 18,
+    width: 13,
+    height: 15,
+    hotspot_x: -2,
+    hotspot_y: -1
+},
+{
+    x: 304,
+    y: 18,
+    width: 14,
+    height: 14,
+    hotspot_x: -2,
+    hotspot_y: 0
+},
+{
+    x: 323,
+    y: 18,
+    width: 14,
+    height: 16,
+    hotspot_x: -2,
+    hotspot_y: 2
+},
+{
+    x: 342,
+    y: 18,
+    width: 13,
+    height: 16,
+    hotspot_x: -3,
+    hotspot_y: 2
+},
+{
+    x: 361,
+    y: 18,
+    width: 14,
+    height: 14,
+    hotspot_x: -3,
+    hotspot_y: -2
+},
+{
+    x: 380,
+    y: 18,
+    width: 15,
+    height: 14,
+    hotspot_x: -1,
+    hotspot_y: 0
+},
+{
+    x: 0,
+    y: 36,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 19,
+    y: 36,
+    width: 17,
+    height: 16,
+    hotspot_x: 1,
+    hotspot_y: 0
+},
+{
+    x: 38,
+    y: 36,
+    width: 17,
+    height: 13,
+    hotspot_x: -1,
+    hotspot_y: -3
+},
+{
+    x: 57,
+    y: 36,
+    width: 13,
+    height: 15,
+    hotspot_x: -1,
+    hotspot_y: -1
+},
+{
+    x: 76,
+    y: 36,
+    width: 14,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 95,
+    y: 36,
+    width: 14,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 114,
+    y: 36,
+    width: 13,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 133,
+    y: 36,
+    width: 14,
+    height: 14,
+    hotspot_x: 1,
+    hotspot_y: -1
+},
+{
+    x: 152,
+    y: 36,
+    width: 15,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 1
+},
+{
+    x: 171,
+    y: 36,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 190,
+    y: 36,
+    width: 17,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 209,
+    y: 36,
+    width: 17,
+    height: 13,
+    hotspot_x: 2,
+    hotspot_y: -4
+},
+{
+    x: 228,
+    y: 36,
+    width: 13,
+    height: 15,
+    hotspot_x: -2,
+    hotspot_y: -1
+},
+{
+    x: 247,
+    y: 36,
+    width: 14,
+    height: 14,
+    hotspot_x: -2,
+    hotspot_y: -1
+},
+{
+    x: 266,
+    y: 36,
+    width: 14,
+    height: 16,
+    hotspot_x: -2,
+    hotspot_y: 2
+},
+{
+    x: 285,
+    y: 36,
+    width: 13,
+    height: 16,
+    hotspot_x: -3,
+    hotspot_y: 2
+},
+{
+    x: 304,
+    y: 36,
+    width: 14,
+    height: 14,
+    hotspot_x: -3,
+    hotspot_y: -2
+},
+{
+    x: 323,
+    y: 36,
+    width: 15,
+    height: 14,
+    hotspot_x: -1,
+    hotspot_y: 0
+},
+{
+    x: 342,
+    y: 36,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 361,
+    y: 36,
+    width: 17,
+    height: 16,
+    hotspot_x: 1,
+    hotspot_y: 0
+},
+{
+    x: 380,
+    y: 36,
+    width: 17,
+    height: 13,
+    hotspot_x: -1,
+    hotspot_y: -3
+},
+{
+    x: 0,
+    y: 54,
+    width: 13,
+    height: 15,
+    hotspot_x: -1,
+    hotspot_y: -1
+},
+{
+    x: 19,
+    y: 54,
+    width: 14,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 0
+},
+{
+    x: 38,
+    y: 54,
+    width: 14,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 57,
+    y: 54,
+    width: 13,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 76,
+    y: 54,
+    width: 14,
+    height: 14,
+    hotspot_x: 1,
+    hotspot_y: -1
+},
+{
+    x: 95,
+    y: 54,
+    width: 15,
+    height: 14,
+    hotspot_x: 0,
+    hotspot_y: 1
+},
+{
+    x: 114,
+    y: 54,
+    width: 16,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 133,
+    y: 54,
+    width: 17,
+    height: 16,
+    hotspot_x: 0,
+    hotspot_y: 2
+},
+{
+    x: 152,
+    y: 54,
+    width: 17,
+    height: 13,
+    hotspot_x: 2,
+    hotspot_y: -4
+}
+];
